[added] extract classes to individual files
